This is a fix for the decompiler.
Currently, we are setting the children types of the
Vallang constructors. There is still a limitation:
we are not setting the most correct type for the
list fields. Instead of defining a list of the base
type, we are defining a list of a generic valueType.
Nonetheless, the test cases that motivate this patch
are passing.